% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-phones.R
\name{align_phones}
\alias{align_phones}
\title{Align two sequences of phonemes using global alignment}
\usage{
align_phones(a, b, fun_match = phone_match_exact, indel = -1)
}
\arguments{
\item{a, b}{two sequences of phonemes to align}

\item{fun_match}{function to compute the similarity scores for each phoneme}

\item{indel}{score (penalty) to assign to an insertion or deletion (indel)}
}
\value{
a list with information about the matching
}
\description{
Align two sequences of phonemes using global alignment
}
