% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignphone-package.R
\docType{package}
\name{alignphone-package}
\alias{alignphone}
\alias{alignphone-package}
\title{alignphone: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@wisc.edu} (0000-0002-8890-5116)

}
\keyword{internal}
