% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-phones.R
\name{align_phones}
\alias{align_phones}
\title{Align two sequences of phonemes using global alignment}
\usage{
align_phones(
  a,
  b,
  fun_match = phone_match_exact,
  indel_create = -2,
  indel_extend = -1
)
}
\arguments{
\item{a, b}{two sequences of phonemes to align}

\item{fun_match}{function to compute the similarity scores for each phoneme}

\item{indel_create}{penalty for creating a gap}

\item{indel_extend}{penalty for extending a gap by one step}
}
\value{
a list with the class \code{phone_alignment} with information about the
matching
}
\description{
The algorithm used here is based on the \href{https://en.wikipedia.org/wiki/Needleman\%E2\%80\%93Wunsch_algorithm}{Needleman-Wunsch algorithm}.
}
\examples{
# Kiss the sky and kiss this guy
a <- c("k", "I", "s", "dh", "4", "s", "k", "@I") |> wiscbet_to_ipa()
b <- c("k", "I", "s", "dh", "I", "s", "g", "@I") |> wiscbet_to_ipa()
align_phones(a, b)
}
